requirements.txt
# ================================================================================
# REQUISITOS DE ZEROX - Sistema de Trading Automatizado con IA
# ================================================================================
# Para instalar todo: pip install -r requirements.txt

# Core - Inteligencia Artificial y Machine Learning
torch>=2.0.0              # Framework de deep learning para la IA
transformers>=4.30.0      # Modelos de lenguaje avanzados
scikit-learn>=1.3.0       # Algoritmos de machine learning clásicos
pandas>=2.0.0             # Manipulación de datos financieros
numpy>=1.24.0             # Cálculos numéricos eficientes
tensorflow>=2.13.0        # Alternativa/complemento a PyTorch

# Trading y Análisis Técnico
ccxt>=4.0.0               # Conexión universal a exchanges (incluye Bitget)
ta>=0.10.2                # Indicadores técnicos (RSI, MACD, etc.)
yfinance>=0.2.28          # Datos históricos de mercado
python-binance>=1.0.17    # API específica de Binance (referencia)
websocket-client>=1.6.1   # Conexión en tiempo real al mercado

# Interfaz Gráfica
PyQt6>=6.5.0              # Framework para la interfaz gráfica
pyqtgraph>=0.13.3         # Gráficos financieros en tiempo real
matplotlib>=3.7.0         # Gráficos adicionales y análisis
plotly>=5.15.0            # Gráficos interactivos avanzados

# Procesamiento de Lenguaje Natural
nltk>=3.8                 # Análisis de sentimiento en noticias
textblob>=0.17.1          # Procesamiento simple de texto
newspaper3k>=0.2.8        # Extracción de noticias
ebooklib>=0.18            # Lectura de libros EPUB
PyPDF2>=3.0.0             # Lectura de PDFs (libros adicionales)

# Base de Datos y Almacenamiento
sqlalchemy>=2.0.0         # ORM para base de datos
# sqlite3 ya viene incluido con Python, no necesita instalarse
redis>=4.6.0              # Cache de alta velocidad
h5py>=3.9.0               # Almacenamiento eficiente de datos

# Utilidades y Seguridad
python-dotenv>=1.0.0      # Variables de entorno
cryptography>=41.0.0      # Encriptación de API keys
requests>=2.31.0          # Peticiones HTTP
aiohttp>=3.8.5            # Peticiones asíncronas
schedule>=1.2.0           # Programación de tareas
colorama>=0.4.6           # Colores en terminal
tqdm>=4.65.0              # Barras de progreso

# Análisis y Optimización
scipy>=1.11.0             # Optimización matemática
statsmodels>=0.14.0       # Modelos estadísticos
arch>=6.1.0               # Modelos GARCH para volatilidad
hmmlearn>=0.3.0           # Modelos ocultos de Markov

# Testing y Desarrollo
pytest>=7.4.0             # Framework de testing
pytest-asyncio>=0.21.0    # Testing asíncrono
black>=23.7.0             # Formateo de código
pylint>=2.17.0            # Análisis de código

# Compilación a EXE
pyinstaller>=5.13.0       # Crear ejecutable .exe
auto-py-to-exe>=2.36.0    # GUI para PyInstaller

# APIs y Comunicación
python-telegram-bot>=20.4 # Notificaciones por Telegram (opcional)
discord.py>=2.3.0         # Notificaciones por Discord (opcional)
twilio>=8.5.0             # SMS de alertas (opcional)

# Monitoreo y Logs
loguru>=0.7.0             # Sistema de logging avanzado
sentry-sdk>=1.29.0        # Monitoreo de errores en producción
prometheus-client>=0.17.0 # Métricas del sistema

# Optimización de Rendimiento
numba>=0.57.0             # Compilación JIT para velocidad
cython>=3.0.0             # Optimización de código crítico
joblib>=1.3.0             # Paralelización de tareas

# Versiones específicas para estabilidad
Pillow>=10.0.0            # Procesamiento de imágenes
python-dateutil>=2.8.2    # Manejo de fechas
pytz>=2023.3              # Zonas horarias
six>=1.16.0               # Compatibilidad Python 2/3